Moved config to SPIFFS, added pages to serve to SPIFFS, added comments
Now ESP8266 is using SPI Flash File System. It storages config and pages to serve client. This version of program is using ArduinoJson library. Config file contains SSID and PSK to WiFi, so after flashing this file config mode won't start and ESP8266 will connect to WiFi defined in this file. It also contains controller address and device id (idx), which is using in domoticz, and OTA hostname and password (plain text).